+++ src/main/java/com/shopping/cart/repository/ProductRepository.java
@@ -1,5 +1,7 @@
 package com.shopping.cart.repository;
 
+import java.util.Optional;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -10,6 +12,8 @@ import com.shopping.cart.entity.Product;
  *
  */
 @Repository
-public interface ProductRepository extends JpaRepository<Product, Long> {
+public interface ProductRepository extends JpaRepository<Product, Integer> {
 
+//	Optional<Product> findByName(String name);
+//	Optional<Product> findByCategory(String category);
 }
